create a NavLinks component
