+++ src/components/NavigationLink.jsx
@@ -7,7 +7,7 @@ export function NavigationLink({ text, destination, handleClick, icon }) {
 			to={destination}
 			className={({ isActive }) =>
 				[
-					'flex flex-row lg:pr-12 lg: pt-0 pt-6 font-poppins text-lg',
+					'flex flex-row lg:pr-12 lg:pt-0 pt-6 font-poppins text-lg',
 					isActive ? 'font-bold underline' : '',
 				].join(' ')
 			}
