+++ src/components/index.js
@@ -6,3 +6,4 @@ export * from './Loading';
 export * from './NavBar';
 export * from './NavigationLink';
 export * from './ButtonWithIcon';
+export * from './NavLinks';
